example_project/requirements.txt
redis
django-celery
